fix: Add AppDelegate and SceneDelegate, and fix Info.plist
The delegates are an important part of the app life cycle. The choice of using SceneDelegate instead of just AppDelegate is to follow the most recent Apple guidelines regarding the app's life cycle. This commit also fixes the Info.plist file because it had a reference to an inexistent storyboard.